Defined all Metasprite data (#5)
* Defined all Metasprite Data

I got bored and wanted to finish since I need this for my project, so a lot is defined as "Unknown", or "Part2Unknown" (Since I started on the second half). Should be much easier to go through now if one needs to edit the metasprites, should save someone the days it took me in the future.